Readline algorithm has some issues
